+++ .ts-for-gir.packages-all.rc.js
@@ -38,8 +38,8 @@ export default {
         // GNOME Shell gir file dependencies on Ubuntu 23.10 (package: libmutter-13-dev)
         '/usr/lib/x86_64-linux-gnu/mutter-13',
 
-        // Wait for release...
-        '/usr/lib/x86_64-linux-gnu/mutter-14',
+        // GNOME Shell gir file dependencies on Ubuntu package: libmutter-14-dev) and GNOME OS
+        '/usr/lib/x86_64-linux-gnu/mutter-14', 
 
         // General gir files in this repository
         './vala-girs/gir-1.0',